Gitlab connection (#49)
* Make `docker compose` happy.

* Fix dependencies in Makefile.

Many of the targets required a `make clean` before they would actually
do work if the tree had ever been initialized.  I pulled from upstream
and was surprised that `make setup` didn't do anything but install
packages and bootstrap even though the DDL had been modified.

* Add missing files to ci/test copy.

I added some dependencies in 145c217b that need to be present in the CI
build as well.  I also switched to using tar instead of cp since it will
do the job more quickly.

* Use venv in tests too.

Instead of doing a pip install and assuming that it went to /usr/local,
just use a virtual env.  It adds nearly zero overhead and unifies the
dev and CI environments.

* Add integration columns to projects.

* Describe how to add new user in CONTRIBUTING.

* Add RO CRUD endpoints for integrated applications.

* Add "integrations" to the user model.

* Implement connecting to gitlab.

* Clean up logging configuration.

The gist of this is to register the handler using "root" so that output
from unconfigured handlers is actually logged.

* Add POST /integrations

* Add form encoding support for POST /integrations.

This just makes things much easier when using curl.

* Document gitlab integration in CONTRIBUTING.

* Add DELETE /integrations/{name}

* Make SimpleForm button text configurable.

* Implement project import from gitlab.

Lots of things are questionable in this commit... I'm not a JS
programmer but I occasionally think about maybe playing one on TV in
some future life.

* Update project fields from gitlab info.

I was previously stuffing data into hard-coded project facts.  These
have been replaced by the new columns in the projects table.

* Fix flake8 errors that went undetected.