Create gitlab project (#56)
* Actually log requests that return 500

* Add openapi spec for /gitlab endpoints.

* Pull GitLabClient out of handlers.

This simplifies the endpoints significantly, reduces duplication, and
opens us up to implementing auto token refreshing in the future.

* Add namespaces.gitlab_group_id.

* Add project_types.gitlab_project_prefix.

* Add gitlab project creation endpoint.

I popped this under /ui/automations/... since it isn't generally useful
unless some one is logged in.

* Create gitlab repo during project creation.

I'm not entirely pleased with out this turned out since you can enable
the "create gitlab repo" toggle then change the project to something that
does not have a gitlab prefix and the result is that the toggle is
enabled and will be ignored :/

I could not find a good way to call `setAutomations` and have the result
reflected in the state of the UI element.

* Add errors.py.

This will be used to make error response documents more consistent
throughout the application.

* Add canonical_server_name setting.

This is used to configure the server name that is used in the error
type URLs.

* Weave errors.py usage through application.

* Remove unnecessary params in gitlab creation.

Turns out that the imbi project already contains the things that I need
to create the gitlab project so only the imbi project ID is required.

* Create "automation" classes.

This commit moves the core of the GitLabCreateProjectAutomation
into a new module that will provide a home for other automations
in addition to any configuration that they may need.  It seemed odd to
bury this stuff in the guts of the UI endpoints.

* Remove gitlab config sub-tree.

The application details are configured when you create the GitLab
OAuth 2 integration by POSTing to /integrations.  The client id and
secret are stored in the `oauth_integrations` table.

* Make GitLab automation create project link.

This needs to be configured in the `automations` sub-tree of the
configuration file.  If it is not configured, then no link will be
created in the project.