Fix tests.
The namespaces and project-types endpoint tests were correctly detecting
changes that were made in the previous PR ... I forgot to update the
tests before creating it.  This updates the tests to be more resilient
by using the class-level metadata that we have to create models with
all of the proper fields populated correctly.

The tests will still fail after this commit since the OpenAPI
specification for the project-types/{id} endpoints is incorrect.  That
will be fixed in the next commit.