Fix internal user passwords (#57)
* Add GitLab group name to Namespace admin UI.

* Add GitLab prefix to Project Type admin UI.

* Typing adjustments for cross imports.

I ran the tests with pytest and it had a number of problems related to
circular imports and compile time type annotations.  I had to change a
number of them to forward referencing (quoted) annotations.

* Fix tests.

The namespaces and project-types endpoint tests were correctly detecting
changes that were made in the previous PR ... I forgot to update the
tests before creating it.  This updates the tests to be more resilient
by using the class-level metadata that we have to create models with
all of the proper fields populated correctly.

The tests will still fail after this commit since the OpenAPI
specification for the project-types/{id} endpoints is incorrect.  That
will be fixed in the next commit.

* Fix OpenAPI spec for project-types.

The gitlab_project_prefix should have been nullable.

* Add keychain module.

This will replace the usage of the tornado secure cookie stuff for
passwords.  Secure cookies include a timestamp so the following sequence
will always fail:

  encrypted = self.application.encrypt_value('password', 'foo')
  sleep(1)
  assert self.application.encrypt_value('password', 'foo') == encrypted

This is likely what is causing the seemingly random test failures
centered around the authorization of internal users.

* Add encryption_key application setting.

* Switch from storing encrypted passwds to hashes.

I also switched the encrypted passwords stored in the session to using
cryptography.Fernet instead of tornado's cookie secret.

* Remove now unused params to encrypt/decrypt.

* Insert hash algorithm in hashed password values.

* Switch to openssl for the time being.

There is a known issue with Alpine 3.14 and libressl-dev.  I also had to
add cargo to the image for the Cryptography build to work.

* Add "canonical_server_name" to test config.