Add first commit from a cookiecutter (#59)
* Add some missing UI labels.

* Move project lookup into automations.

This was previously done in the UI endpoint but we are going to need a
lot more data for the cookiecutter so I decided to mvoe it into the
automations code proper.

* Implement first commit to git from a cookiecutter.

* Populate the Sonar link if configured to do so.

Most of the cookiecutters conditionally include sonar settings, an
analysis stage, and links if the sonar properties are configured.

* Run yapf & isort over template expansion.

These are required to ensure that the code meets the expectations before
we commit it and execute the CI pipeline.

* Be more defensive about GitLab settings.

* Pass the application through to automations.

* Make yapf & isort configurable.

* Fix an accidental break in GitLab importing.

* Implement gitlab.restrict_to_user for development.

This makes it safer since we will run this way locally and force newly
created GitLab projects to be created in the user's namespace by
default.  You can always manually toggle the setting in debug.yaml if
you want to spray projects all over GitLab.

* Add short_description to GitLab projects.

Also fix a import order sorting violation.

* Describe SonarQube project creation in decisions.

* Use pydantic for GitLab response decoding.

This makes everything so much easier.

* Fix yapf formatting.

The configuration wasn't being passed along :/

* Create SonarQube project after GitLab project.

I had to do some interesting work in the Create view so that we trigger
the SonarQube automation between the GitLabCreateProjectAutomation and
the GitLabInitialCommitAutomation since we need the SonarQube project to
exist *before* we commit to GitLab.  Otherwise the initial Sonar
analysis will be broken as described in decisions/0011-sonarqube-projects.md

* Enable MR decoration if GitLab ALM is configured.

This is not reliably enabled for newly created projects so we might as
well enable it at project creation time.

* Add Imbi project link to SonarQube dashboard.

* Remove ambiguity in project_fact OpenAPI def.

POST to create new project facts would fail because the schema for the
fact value was ambiguous when the value was a string.  The OpenAPI
validation code could not decide between string, date, or
iso8601-timestamp so it would give up and raise an exception.

* Refactor sonarqube client.

I broke the functional pieces into separate async methods that could
then be called in parallel.

* Extract GitLabClient from endpoints.gitlab.

* Add search for projects by sonarqube_project_key.

This is going to be necessary for updating projects based on SonarQube
webhook notifications.