Fix operations-log endpoint response ordering
This fixes a few bugs unearthed while testing. The first is when
following previous links, the ordering of asc/desc and the anchor
means that you ended up grabbing rows either at the very beginning
or very end of the collection. So during a previous link, the
ordering in the db query is reversed to grab the data closest to the
anchor, and then the resulting rows are then reversed again to revert
back to the original requested ordering.

The second is when previous and next links are required. It handles
all cases when all limit + 1 entries were found, indicating that there
is another page, either a next page if you're following a next page,
or vice-versa if you're following a previous page. Then, regardless,
if you are following a next link there will be a previous page,
and if following a previous link there will be a next page.