Implement input sanitization for project facts
I reworked the common.coerce_project_fact_values function to perform
more rigorous input sanitization, changed `POST /project/{id}/facts`
to call it to validate the request body, and added tests for the whole
mess.

This changes the response code from a 500 to a 400 when it is given
anything completely incorrect.  I was pretty permissive on the input
side of things so stringified numbers will be coerced correctly based on
the fact type and I used dateutil.parser.parse for the date and
timestamp parsing since it behaves considerably better than arrow does.

coerce_project_fact_values is used when fetching project facts as well
so the result is that we get very consistent output.