Stop setting the platform for Sentry projects
We definitely don't want to hard code this to Python so the value will
have to be dynamically determined from some project fact.  However, I
think that the sentry client will overwrite this for us so I am going to
leave a TODO in there for the time being and create an issue to revisit
after we have some insight from actual usage.

https://github.com/AWeber-Imbi/imbi/issues/73