Add ui endpoint for Google login
This is the API side of moving the redirect to Google's authorization
endpoint from the frontend to the backend. By hitting the backend
first, we add a state token to the request query param, which acts
as a CSRF token and will be validated when Google returns to our
auth endpoint.