Use new upsert_user postgres function
We need to handle the case where there is a conflict on email_address,
not just username, as Google users may have the same email as their
LDAP counterparts, and we wish to update that user. Since we can't
do two separate ON CONFLICT sections, we moved the logic into a
stored proc.