Make user.groups a list of strings
This deduplicates the group permissions data (which was stored in
both the permissions of each Group object and also the permissions
field itself). This also fixes a bug where we expected Group objects
in the groups field, but when retrieving the data from Redis, we only
had the list of strings. This just makes it a list of strings from
end-to-end, making it less complicated.