Use list of strings for user integrations
Similar to the previous commit, simplify the user object connected
integrations to be just a list of strings for each integration.
This fixes a similar bug with the fetch from Redis. This does lose
us the integration ID, but we don't use it (YAGNI - and if we do
need it later we can still get it since we get the data from the db).