Replace /integrations and /integrations/{slug}
These were unused from the FE so I am replacing them with endpoints for
retrieving and manipulating content from the `v1.integrations` table.
I apologize for the temporary addition of the `_apps` module. I promise
that I will clean that up shortly.

This commit also adds the ability to specifically omit fields in the
record responses since we do not want to return the integration secret
in the response from creating an integration.