Add POST /integrations/{n}/automations
This is a pretty complex commit ... I'm not sure what the best way to
represent the information here is. We have:

1. database enum type adaptation
2. a new module with slugification logic
3. pydantic request and response models
4. helpers to manage mapping between ids and slugs
5. finally, the request handler code

The mapping helpers make it possible to reference other entities by
either their IDs or slugs which makes for a really nice API. The
`_map_ids_and_slugs` function may actually be useful elsewhere.
Anyway, there is a lot of new stuff going on in this commit that is
foundational for the remaining work.

The slugification helpers may also be useful elsewhere... seems that we
have to have that logic somewhere.