Move "do_nothing" function into imbi.automations
This is going to be the home for embedded automation code. I am also
going to add an allow list that limits where you can run code from. We
don't want an ne'er-do-wells running `sys.exit` do we?

Note that I took the annotations route since that is what Pydantic is
suggesting as the preferred mechanism for type-based constraints as
opposed to field-based constraints. This is definitely the former.