Implement GitLab project creation
This is in the imbi.automations namespace since that is where our allow
listed functionality lives. The creation code itself is straight
forward. I do need to figure out how to isolate the SQL code to create
identifiers and links... I'm going to live with the technical debt for
the time being since we need a more functional model layer to really
address the problem.