Rework how GitLab & Sentry clients are initialized
Created an async ``create_client`` function in both modules
since the configuration is split between the config file and
the database. The create functions will raise an error that
is a sub-class of errors.ApplicationError so users are not
required to do special error handling or transformation.

Note that I did not move the SonarQube client yet. That will
come in a future commit.