Add base.PaginatedCollectionHandler
This is a more generic handler than the TimeBasedPaginationMixin;
however, I don't think that we can implement the latter in terms of the
next handler due to it's initialization requirements. I commented the
new handler and associated token thoroughly. The one wonky part is the
handling of `self.ID_KEY` since it is either None, a string, or an
iterable and we need to deal with it to properly use `reverse_url`.

Note that I switched from `typing` to `typing-extensions` to give me
access to `typing.Self`.