Add /components and /components/{package-url}
The /components endpoint uses token-based pagination using the package
URL as the sort key. If you were taken aback by the complexity in the
previous commit, this commit might convince you why it was necessary.

I had to add a "check validity" hook to the PATCH processing after I
discovered that our OpenAPI-based validation does not help one bit with
validating the updated representations. I used a pydantic model to
validate the new object before it is written to the DB.

I also started carefully separating the perms required for the various
CRUD operations. For example, any authenticated user can issue a
GET /components but only an admin can create new components. This set of
endpoints should serve as an example for making read access to much of
the Admin API available to non-admin users.