Fix component_versions caching during SBoM ingest
We were leaving the empty sets in the cache for packages that were not
in the database. This caused the `KeyError` logic in
`insert_component_version` to fail. I switched to only caching positive
entries instead of priming the cache with every package in the SBoM.
The result is that the cache faithfully represents what is in the
database when we start the ingest.