Fix missing/circular import of imbi.models
As-is, imbi.models is not defined, because we only import it
if typing.TYPE_CHECKING. However, if we import it at the top level
there is a circular import because it tries to import the function
do_nothing from imbi.automations which is not yet defined. There
are a few ways to solve this, such as moving do_nothing into
automations.models since that's the only place it's used. But it
makes sense to keep that function in imbi.automations so instead
I opted here to just import the package in the function when we
need it.