Convert project component scores to numeric values
This change removes the requirement that the project fact value has enum
values that match what is in the code. After some consideration I
decided that requiring a user to manually configure a project fact with
enumerated values that are fixed was just too strange. Now the project
component score is a numeric range with three slots:

 0 - 20 --> unacceptable (red)
20 - 80 --> needs work   (orange)
80 - 100 -> okay         (green)

The next commit will create the project fact type with an appropriate
fact type range definition.