Create component score project fact type in init
If component scoring is enabled in the configuration file, then
auto-create the project fact type if it does not already exist.
There is some interesting logic in app.py that auto-creates the
project fact type that warrants some review.

The range values for the project fact will seem *odd* at first. There
are some clarifying comments in there as well but they match the
expectations set out in the previous commit.