Move scoring initialization into --initialize
This avoids a race condition caused when we run more than one process in
a container :/

I added a flag to the Application to avoid some log messages if we are
running with `--initialize`. We may want to use it to control other
things like not setting up the sentry connection for CLI usage.