Support OpsLog entries for nonexistent users
We supported this at the database level; however, the API would always
silently drop user names that were not in `v1.users`. This commit
changes the queries to return `UNKNOWN` for the email address or display
name when the user does not exist in the database. It is interesting
that we would insert the row into postgres and then fail to retrieve it
and return a 404 since the GET method was using a full join :/