Parse CEL document for CEL evaluation
CEL requires you to define the shape of the "document" that you are
evaluating against the expression. celpy provides a convenience
method to automatically parse it in a best effort. So we'll get
everything like str, bool, int, float, tuple, and mapping for free.
Ideally we don't want to have to define the CEL declarations for each
type of object and we keep this generic like so. This should cover
the vast majority of our use cases.